move decomposition tests to test directory
